moved factories to processors directory
