+++ src/Microsoft.Health.Fhir.Liquid.Converter.UnitTests/Factories/ConvertProcessorFactoryTests.cs
@@ -7,7 +7,6 @@ using System;
 using System.Collections.Generic;
 using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Logging.Abstractions;
-using Microsoft.Health.Fhir.Liquid.Converter.Factories;
 using Microsoft.Health.Fhir.Liquid.Converter.Models;
 using Microsoft.Health.Fhir.Liquid.Converter.Processors;
 using Xunit;
