+++ src/Microsoft.Health.Fhir.Liquid.Converter/Processors/IConvertProcessorFactory.cs
@@ -6,7 +6,7 @@
 using Microsoft.Health.Fhir.Liquid.Converter.Models;
 using Microsoft.Health.Fhir.Liquid.Converter.Processors;
 
-namespace Microsoft.Health.Fhir.Liquid.Converter.Factories
+namespace Microsoft.Health.Fhir.Liquid.Converter.Processors
 {
     public interface IConvertProcessorFactory
     {
